Fix flake8 and amazon unit test
